backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client
openai
pypdf2
python-dotenv
